# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /user/users:
    put:
      tags:
      - user-controller
      summary: Update user account of currently logged-in user.
      description: |
        Update user account of currently logged-in user. Only sub-set of all fields can be updated with this endpoint.
        These fields contain every field which is safe to be changed by regular user.
      operationId: updateWithUserPrivilege
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequestWithUserPrivilege'
        required: true
      responses:
        "404":
          description: Requested User was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "400":
          description: Zero fields were set in update object. At least single filed
            should be set.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "200":
          description: User account was updated successfully.
        "409":
          description: At least one set field which should be unique is not.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
      - Bearer Authentication: []
  /admin/users/{userId}:
    put:
      tags:
      - user-controller
      summary: Update user account with admin privilege.
      description: Update user account with admin privilege. This endpoint requires
        at least admin privilege. All users accounts and all fields can be modified
        with this endpoint.
      operationId: updateWithAdminPrivilege
      parameters:
      - name: userId
        in: path
        description: Id of the user on which specified operation is being performed.
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequestWithAdminPrivilegeDoc'
        required: true
      responses:
        "404":
          description: Requested User was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "400":
          description: Zero fields were set in update object. At least single filed
            should be set.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "200":
          description: User account was updated successfully.
        "409":
          description: At least one set field which should be unique is not.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
      - Bearer Authentication: []
...
components:
  schemas:
    UserUpdateRequestWithUserPrivilege:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
      description: |
        Json with data for performing user update. Each field is optional, only set fields will be used for update.
        This objects contains fields which are safe to be modified with user privilege.
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
    UserUpdateRequestWithAdminPrivilegeDoc:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
          - ADMIN
          - MOD
          - USER
      description: |
        Json with data for performing user update. Each field is optional, only set fields will be used for update.
        This objects contains fields which can be modified with user and admin privilege, thus to use it, one must
        have admin privilege.
...
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT